Good as it Gets
This is as good as it’s going to get for now. Rotation is now accurate
enough that a human can rectify the rotation and roll can be handled by
software. But translation is still not able to be handled by software
though there is the start to some filter code for accelerometer.